npm: Make DetectManifest return the manifest info
Since npm's package-lock.json could be either v1 or v2 and
node_modules/.package-lock.json could be either v2 or v3, the
current manifest detection logic is not enough. We have to further
read the manifest data and, from the data, figure out what
version of the lockfile we are using. Hence we return the type
of manifest.

The current plugin data contract wants the name of the manifest file
as well. To satisfy this, we will also return the manifest file.

- Make the manifest list into a manifest map with the key as the
  possible file path and the value the type of manifest.
- Update the DetectManifest function to return both manifest file and
  manifest type.
- Update tests to reflect the change.

Signed-off-by: Nisha Kumar <nisha.kumar@oracle.com>